Fix some Bash issues

+++ lib/common.sh
@@ -81,7 +81,7 @@ cache_copy() {
   rel_dir=$1
   from_dir=$2
   to_dir=$3
-  rm -rf $to_dir/$rel_dir
+  rm -rf "${to_dir:?}/${rel_dir:?}"
   if [ -d $from_dir/$rel_dir ]; then
     mkdir -p $to_dir/$rel_dir
     cp -pr $from_dir/$rel_dir/. $to_dir/$rel_dir
